added caro game homework with flawed algorithm
currently what I implemented in the checkWin function is that it
will check all the nearby tiles that player has marked. That means if
the player line has an L shape, the algorithm still see it as a win.

Also, if the player put their mark in the middle of the winning line,
the algorithm cannot recognized that